+++ bridge.h
@@ -3,7 +3,7 @@
 
 uint32_t bridge_releaseProcessor(void);
 uint32_t bridge_setProcessPriority(uint8_t pid, uint8_t priority);
-uint8_t bridge_getProcessPriority(uint8_t pid);
+int16_t bridge_getProcessPriority(uint8_t pid);
 
 void *bridge_acquireMemoryBlock(void);
 int8_t bridge_releaseMemoryBlock(void *blk);
